+++ packages/terra-clinical-header/tests/nightwatch/HeaderTestRoutes.jsx
@@ -10,6 +10,7 @@ import RightContentHeader from './RightContentHeader';
 import LeftContentHeader from './LeftContentHeader';
 import LeftAndRightContentHeader from './LeftAndRightContentHeader';
 import LongTitleAndContentHeader from './LongTitleAndContentHeader';
+import LeftAndRightContentSubheader from './LeftAndRightContentSubheader';
 
 const routes = (
   <div>
@@ -21,6 +22,7 @@ const routes = (
     <Route path="/tests/header-tests/left-content" component={LeftContentHeader} />
     <Route path="/tests/header-tests/left-and-right-content" component={LeftAndRightContentHeader} />
     <Route path="/tests/header-tests/long-title-content" component={LongTitleAndContentHeader} />
+    <Route path="/tests/header-tests/left-and-right-content-subheader" component={LeftAndRightContentSubheader} />
   </div>
 );