+++ packages/terra-clinical-site/CHANGELOG.md
@@ -22,6 +22,9 @@ Unreleased
 ### Changed
 * Update Header example
 
+### Added
+* Example to render subheader component
+
 1.1.0 - (July 18, 2017)
 -----------------
 ### Changed